+++ site/src/components/SSHButton/SSHButton.stories.tsx
@@ -1,8 +1,5 @@
 import { Story } from "@storybook/react"
-import {
-  MockWorkspace,
-  MockWorkspaceAgent,
-} from "../../testHelpers/renderHelpers"
+import { MockWorkspace, MockWorkspaceAgent } from "testHelpers/entities"
 import { SSHButton, SSHButtonProps } from "./SSHButton"
 
 export default {
