+++ site/src/components/UsersTable/UsersTable.stories.tsx
@@ -1,9 +1,9 @@
 import { ComponentMeta, Story } from "@storybook/react"
 import {
-  MockAssignableSiteRoles,
   MockUser,
   MockUser2,
-} from "../../testHelpers/renderHelpers"
+  MockAssignableSiteRoles,
+} from "testHelpers/entities"
 import { UsersTable, UsersTableProps } from "./UsersTable"
 
 export default {
